Merge existing library variables when importing
When importing an updated library variable set into a different
Octopus server, the server might contain additional variables or
scoped values that do not appear in the import file.  To avoid
destroying environment-specific variable values when importing,
existing variables that don't appear in the import with an
equivalent scope are merged with the imported variables.  Existing
variables with equivalent scopes will be replaced by the updated
values in the import file.