Renovated the CreateReleaseCommand for --autochannel feature.
In order to select the right channel based on the provided package versions, we need to simulate what it would happen if we created that release into each of the channels, and pick the right one based on which ReleasePlan was the most viable.

This means we needed a way to reliably and consistently calculate the ReleasePlan, instead of doing it inline in the CreateReleaseCommand. This logic has been moved into the ReleasePlanBuilder.

Now the CreateReleaseCommand can create all of the possible ReleasePlan combinations and select the most viable.

I think overall this has simplified the code quite a lot, and consolidated a lot of the responsibilities into the building of ReleasePlan, instead of scattered around the code base.

I've ended up removing these tests temporarily and will add tests to exercise the new shape.