Isolate anti-forgery token by InstallationId
In the rare-ish case where customers have multiple Octopus installations on the same domain (it does happen, quite a lot in our test environments too) we will end up with multiple cookies in the CookieCollection for that domain, and we were looking for the first match based on the common prefix. Quite often this would end up with the wrong anti-forgery cookie for that installation and fail validation. We should use the InstallationId to partition and find the correct anti-forgery cookie.

This should be both backwards and forwards compatible.